+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- German translations to checkout6-custom messages
+
 ## [3.0.2] - 2025-10-14
 
 ### Fixed